Get process names from process codes

Add a new column to an existing tibble with the corresponding name for each process code. The codes are assumed to be from those defined by the FABIO model.

Usage

add_process_name(
  table,
  code_column = "process_code",
  name_column = "process_name"
)

Arguments

table

The table that will be modified with a new column.

code_column

The name of the column in table containing the codes.

name_column

The name of the output column containing the names.

Value

A tibble with all the contents of table and an extra column named name_column, which contains the names. If there is no name match, an NA is included.

Examples

table <- tibble::tibble(process_code = c("p017", "p076", "dummy"))
add_process_name(table)
#> # A tibble: 3 × 2
#>   process_code process_name        
#>   <chr>        <chr>               
#> 1 p017         Beans production    
#> 2 p076         Olive Oil extraction
#> 3 dummy        NA                  

table |>
  dplyr::rename(my_process_code = process_code) |>
  add_process_name(code_column = "my_process_code")
#> # A tibble: 3 × 2
#>   my_process_code process_name        
#>   <chr>           <chr>               
#> 1 p017            Beans production    
#> 2 p076            Olive Oil extraction
#> 3 dummy           NA                  

add_process_name(table, name_column = "my_custom_name")
#> # A tibble: 3 × 2
#>   process_code my_custom_name      
#>   <chr>        <chr>               
#> 1 p017         Beans production    
#> 2 p076         Olive Oil extraction
#> 3 dummy        NA