Primary production items linked to CBS

Maps FAOSTAT primary production items and crop products to their CBS counterparts, along with farm and labour classifications.

Usage

items_prim

Format

A tibble where each row corresponds to one production item. It contains the following columns:

item_prod: Name of the production item (e.g., "Wheat", "Rice").

item_prod_code: FAOSTAT production item code (character).

item_cbs: Name of the corresponding CBS item.

item_cbs_code: Numeric CBS item code.

Farm_class: Farm system classification. Crop items use codes such as "COP" (cereals, oilseeds, protein crops), "Vegetables", "Fruits", "Olive", "Grapevine", "Other_crops". Livestock items use "Dairy_cows", "Cattle", "Monogastric", "Sheep_goats", "Bees", "Game". NA for non-farm items.

Cat_Labour: Labour category used in agricultural labour analyses.

Cat_FAO1: Top-level FAO commodity category.

group: Item group classification. One of "Primary crops", "Crop products", "Livestock products", "Grass", "Crop residues", "Scavenging", "Livestock".

Source

Derived from FAOSTAT Production data.

Examples

head(items_prim)
#> # A tibble: 6 × 8
#>   item_prod item_prod_code item_cbs item_cbs_code Farm_class Cat_Labour Cat_FAO1
#>   <chr>     <chr>          <chr>            <dbl> <chr>      <chr>      <chr>   
#> 1 Wheat     15             Wheat a…          2511 COP        Cereals    Cereals 
#> 2 Rice      27             Rice an…          2807 COP        Rice       Cereals 
#> 3 Barley    44             Barley …          2513 COP        Cereals    Cereals 
#> 4 Beer of … 51             Beer              2656 NA         NA         NA      
#> 5 Maize (c… 56             Maize a…          2514 COP        Cereals    Cereals 
#> 6 Oil of m… 60             Maize G…          2582 NA         NA         NA      
#> # ℹ 1 more variable: group <chr>