Processed products share factors

Clean CSV data from the given file_path.

Usage

get_processing_coefs(file_path)

Arguments

file_path

The local path where the input CSV is located. It is recommended to use an autogenerated path by get_file_path() function.

Value

A tibble with the quantities for each processed product. It contains the following columns:

year: The year in which the recorded event occurred.

area: The name of the country where the data is from.

area_code: FAOSTAT internal code for each country. Equivalences with ISO 3166-1 numeric can be found in the Area Codes CSV from the zip file that can be downloaded from FAOSTAT. TODO: Think about this, would be nice to use ISO3 codes but won't be enough for our periods

item_to_process: Natural language name for the item that is being processed and will give other subproduct items.

item_to_process_code: FAOSTAT internal code for each of those items.

value_to_process: tonnes of this item that are being processed. It matches the amount found in the processing column from the data obtained by get_wide_cbs().

item_processed: Natural language name for the subproduct item obtained.

initial_conversion_factor: estimate for the number of tonnes of item_processed obtained for each tonne of item_to_process. It will be used to compute the final_conversion_factor, which leaves everything balanced. TODO: explain how it's computed.

initial_value_processed: first estimate for the number of tonnes of item_processed obtained from item_to_process. It is computed as value_to_process * initial_conversion_factor.

conversion_factor_scaling: computed scaling needed to adapt initial_conversion_factor so as to get a final balanced total of subproduct quantities. TODO: explain how it's computed.

final_conversion_factor: final used estimate for the number of tonnes of item_processed obtained for each tonne of item_to_process. It is computed as initial_conversion_factor * conversion_factor_scaling.

final_value_processed: final estimate for the number of tonnes of item_processed obtained from item_to_process. It is computed as initial_value_processed * final_conversion_factor.

For the final data obtained, the quantities final_value_processed are balanced in the following sense: the total sum of final_value_processed for each unique tuple of (year, area_code, item_processed) should be exactly the quantity reported for that year, country and item_processed item in the production column obtained from get_wide_cbs(). This is because they are not primary products, so the amount from 'production' is actually the amount of subproduct obtained. TODO: Fix few data where this doesn't hold.

Examples

if (FALSE) { # \dontrun{
get_processing_coefs(get_file_path("processing_coefs"))
} # }